[Implementation] Completed the basic overview of class structure
Years have months, and months have days. Each year will only have 12
months, and each month is set to have the correct number of days. There
are not automatically the max number of days in the month, but we only
hold the days that weight was recorded to save space.